+++ cmd/som/cli/plot_umatrix.go
@@ -19,9 +19,16 @@ func plotUMatrixCommand() *cobra.Command {
 
 	command := &cobra.Command{
 		Use:   "u-matrix [flags] <som-file> <out-file>",
-		Short: "Plots the u-matrix of an SOM, showing inter-nde distances",
-		Long:  `Plots the u-matrix of an SOM, showing inter-nde distances`,
-		Args:  cobra.ExactArgs(2),
+		Short: "Plots the u-matrix of an SOM, showing inter-node distances.",
+		Long: `Plots the u-matrix of an SOM, showing inter-node distances.
+
+Data provided via --data-file can be displayed on top of the u-matrix,
+showing the values in the column given by the --label flag:
+
+  som plot u-matrix som.yml u-matrix.png --data-file data.csv --label name
+
+For large datasets, --sample can be used to show only a sub-set of the data.`,
+		Args: cobra.ExactArgs(2),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			somFile := args[0]
 			outFile := args[1]
@@ -39,9 +46,9 @@ func plotUMatrixCommand() *cobra.Command {
 	}
 
 	command.Flags().IntSliceVarP(&size, "size", "s", []int{600, 400}, "Size of the plot in pixels")
-	command.Flags().StringVarP(&dataFile, "data-file", "f", "", "Data file. Required for --labels")
+	command.Flags().StringVarP(&dataFile, "data-file", "f", "", "Data file. Required for --label")
 	command.Flags().StringSliceVarP(&ignore, "ignore", "i", []string{}, "Ignore these layers for BMU search")
-	command.Flags().StringVarP(&labelsColumn, "labels", "l", "", "Labels column in the data file")
+	command.Flags().StringVarP(&labelsColumn, "label", "l", "", "Label column in the data file")
 	command.Flags().IntVarP(&sample, "sample", "S", 0, "Sample this many rows from the data file (default all)")
 
 	command.Flags().StringVarP(&delim, "delimiter", "D", ",", "CSV delimiter")